requirements.txt
pyserial
tkinter
customtkinter